evm/tx/common/statemanager: move EEI logic into statemanager
