+++ packages/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './common'
 export * from './enums'
+export * from './interfaces'
 export * from './types'
 export * from './utils'
