+++ packages/evm/package.json
@@ -47,17 +47,16 @@
     "tsc": "../../config/cli/ts-compile.sh"
   },
   "dependencies": {
+    "@ethereumjs/statemanager": "^1.0.4",
     "@ethereumjs/common": "^3.1.1",
     "@ethereumjs/util": "^8.0.5",
     "@ethersproject/providers": "^5.7.1",
-    "@ethereumjs/tx": "^4.1.1",
     "debug": "^4.3.3",
     "ethereum-cryptography": "^1.1.2",
     "mcl-wasm": "^0.7.1",
     "rustbn.js": "~0.2.0"
   },
   "devDependencies": {
-    "@ethereumjs/statemanager": "^1.0.4",
     "@ethersproject/abi": "^5.0.12",
     "@types/benchmark": "^1.0.33",
     "@types/core-js": "^2.5.0",
