+++ packages/evm/src/evm.ts
@@ -1,4 +1,5 @@
 import { Chain, Common, Hardfork } from '@ethereumjs/common'
+import { DefaultStateManager } from '@ethereumjs/statemanager'
 import {
   Account,
   Address,
@@ -22,28 +23,24 @@ import { Interpreter } from './interpreter'
 import { Message } from './message'
 import { getOpcodesForHF } from './opcodes'
 import { getActivePrecompiles } from './precompiles'
-import { DefaultBlockchain, EEI } from './state/state'
 import { TransientStorage } from './transientStorage'
+import { DefaultBlockchain } from './types'
 
 import type { InterpreterOpts, RunState } from './interpreter'
 import type { MessageWithTo } from './message'
 import type { OpHandler, OpcodeList } from './opcodes'
 import type { AsyncDynamicGasHandler, SyncDynamicGasHandler } from './opcodes/gas'
 import type { CustomPrecompile, PrecompileFunc } from './precompiles'
-import type { Blockchain } from './state/state'
 import type {
   Block,
+  Blockchain,
   CustomOpcode,
-  EEIInterface,
   EVMEvents,
   EVMInterface,
   EVMRunCallOpts,
   EVMRunCodeOpts,
-  /*ExternalInterface,*/
-  /*ExternalInterfaceFactory,*/
   Log,
 } from './types'
-import type { StateManagerInterface } from '@ethereumjs/common'
 
 const debug = createDebugLogger('evm:evm')
 const debugGas = createDebugLogger('evm:gas')
@@ -142,7 +139,7 @@ export interface EVMOpts {
   /*
    * The StateManager which is used to update the trie
    */
-  stateManager: StateManagerInterface
+  stateManager: DefaultStateManager
 
   /**
    *
@@ -192,7 +189,8 @@ export class EVM implements EVMInterface {
 
   readonly _common: Common
 
-  public eei: EEIInterface
+  public stateManager: DefaultStateManager
+  public blockchain: Blockchain
 
   public readonly _transientStorage: TransientStorage
 
@@ -290,7 +288,8 @@ export class EVM implements EVMInterface {
       blockchain = opts.blockchain
     }
 
-    this.eei = new EEI(opts.stateManager, this._common, blockchain)
+    this.blockchain = blockchain
+    this.stateManager = opts.stateManager ?? new DefaultStateManager()
 
     // Supported EIPs
     const supportedEIPs = [
@@ -375,7 +374,7 @@ export class EVM implements EVMInterface {
   }
 
   protected async _executeCall(message: MessageWithTo): Promise<EVMResult> {
-    let account = await this.eei.getAccount(message.authcallOrigin ?? message.caller)
+    let account = await this.stateManager.getAccount(message.authcallOrigin ?? message.caller)
     if (!account) {
       account = new Account()
     }
@@ -389,7 +388,7 @@ export class EVM implements EVMInterface {
       }
     }
     // Load `to` account
-    let toAccount = await this.eei.getAccount(message.to)
+    let toAccount = await this.stateManager.getAccount(message.to)
     if (!toAccount) {
       toAccount = new Account()
     }
@@ -453,7 +452,7 @@ export class EVM implements EVMInterface {
   }
 
   protected async _executeCreate(message: Message): Promise<EVMResult> {
-    let account = await this.eei.getAccount(message.caller)
+    let account = await this.stateManager.getAccount(message.caller)
     if (!account) {
       account = new Account()
     }
@@ -482,7 +481,7 @@ export class EVM implements EVMInterface {
     if (this.DEBUG) {
       debug(`Generated CREATE contract address ${message.to}`)
     }
-    let toAccount = await this.eei.getAccount(message.to)
+    let toAccount = await this.stateManager.getAccount(message.to)
     if (!toAccount) {
       toAccount = new Account()
     }
@@ -505,8 +504,8 @@ export class EVM implements EVMInterface {
       }
     }
 
-    await this.eei.putAccount(message.to, toAccount)
-    await this.eei.clearContractStorage(message.to)
+    await this.stateManager.putAccount(message.to, toAccount, true)
+    await this.stateManager.clearContractStorage(message.to)
 
     const newContractEvent = {
       address: message.to,
@@ -515,7 +514,7 @@ export class EVM implements EVMInterface {
 
     await this._emit('newContract', newContractEvent)
 
-    toAccount = await this.eei.getAccount(message.to)
+    toAccount = await this.stateManager.getAccount(message.to)
     if (!toAccount) {
       toAccount = new Account()
     }
@@ -648,7 +647,7 @@ export class EVM implements EVMInterface {
       result.returnValue !== undefined &&
       result.returnValue.length !== 0
     ) {
-      await this.eei.putContractCode(message.to, result.returnValue)
+      await this.stateManager.putContractCode(message.to, result.returnValue)
       if (this.DEBUG) {
         debug(`Code saved on new contract creation`)
       }
@@ -659,8 +658,8 @@ export class EVM implements EVMInterface {
         // This contract would be considered "DEAD" in later hard forks.
         // It is thus an unnecessary default item, which we have to save to disk
         // It does change the state root, but it only wastes storage.
-        const account = await this.eei.getAccount(message.to)
-        await this.eei.putAccount(message.to, account ?? new Account())
+        const account = await this.stateManager.getAccount(message.to)
+        await this.stateManager.putAccount(message.to, account ?? new Account(), true)
       }
     }
 
@@ -678,7 +677,7 @@ export class EVM implements EVMInterface {
     message: Message,
     opts: InterpreterOpts = {}
   ): Promise<ExecResult> {
-    let contract = await this.eei.getAccount(message.to ?? Address.zero())
+    let contract = await this.stateManager.getAccount(message.to ?? Address.zero())
     if (!contract) {
       contract = new Account()
     }
@@ -700,7 +699,13 @@ export class EVM implements EVMInterface {
       versionedHashes: message.versionedHashes ?? [],
     }
 
-    const interpreter = new Interpreter(this, this.eei, env, message.gasLimit)
+    const interpreter = new Interpreter(
+      this,
+      this.stateManager,
+      this.blockchain,
+      env,
+      message.gasLimit
+    )
     if (message.selfdestruct) {
       interpreter._result.selfdestruct = message.selfdestruct as { [key: string]: Uint8Array }
     }
@@ -758,14 +763,14 @@ export class EVM implements EVMInterface {
 
       const value = opts.value ?? BigInt(0)
       if (opts.skipBalance === true) {
-        callerAccount = await this.eei.getAccount(caller)
+        callerAccount = await this.stateManager.getAccount(caller)
         if (!callerAccount) {
           callerAccount = new Account()
         }
         if (callerAccount.balance < value) {
           // if skipBalance and balance less than value, set caller balance to `value` to ensure sufficient funds
           callerAccount.balance = value
-          await this.eei.putAccount(caller, callerAccount)
+          await this.stateManager.putAccount(caller, callerAccount, true)
         }
       }
 
@@ -788,13 +793,13 @@ export class EVM implements EVMInterface {
 
     if (message.depth === 0) {
       if (!callerAccount) {
-        callerAccount = await this.eei.getAccount(message.caller)
+        callerAccount = await this.stateManager.getAccount(message.caller)
       }
       if (!callerAccount) {
         callerAccount = new Account()
       }
       callerAccount.nonce++
-      await this.eei.putAccount(message.caller, callerAccount)
+      await this.stateManager.putAccount(message.caller, callerAccount, true)
       if (this.DEBUG) {
         debug(`Update fromAccount (caller) nonce (-> ${callerAccount.nonce}))`)
       }
@@ -804,10 +809,10 @@ export class EVM implements EVMInterface {
 
     if (!message.to && this._common.isActivatedEIP(2929) === true) {
       message.code = message.data
-      this.eei.addWarmedAddress((await this._generateAddress(message)).bytes)
+      this.stateManager.addWarmedAddress((await this._generateAddress(message)).bytes)
     }
 
-    await this.eei.checkpoint()
+    await this.stateManager.checkpoint()
     if (this._common.isActivatedEIP(1153)) this._transientStorage.checkpoint()
     if (this.DEBUG) {
       debug('-'.repeat(100))
@@ -857,13 +862,13 @@ export class EVM implements EVMInterface {
       !(this._common.hardfork() === Hardfork.Chainstart && err.error === ERROR.CODESTORE_OUT_OF_GAS)
     ) {
       result.execResult.logs = []
-      await this.eei.revert()
+      await this.stateManager.revert()
       if (this._common.isActivatedEIP(1153)) this._transientStorage.revert()
       if (this.DEBUG) {
         debug(`message checkpoint reverted`)
       }
     } else {
-      await this.eei.commit()
+      await this.stateManager.commit()
       if (this._common.isActivatedEIP(1153)) this._transientStorage.commit()
       if (this.DEBUG) {
         debug(`message checkpoint committed`)
@@ -940,7 +945,7 @@ export class EVM implements EVMInterface {
         message.code = precompile
         message.isCompiled = true
       } else {
-        message.containerCode = await this.eei.getContractCode(message.codeAddress)
+        message.containerCode = await this.stateManager.getContractCode(message.codeAddress)
         message.isCompiled = false
         if (this._common.isActivatedEIP(3540)) {
           message.code = getEOFCode(message.containerCode)
@@ -956,7 +961,7 @@ export class EVM implements EVMInterface {
     if (message.salt) {
       addr = generateAddress2(message.caller.bytes, message.salt, message.code as Uint8Array)
     } else {
-      let acc = await this.eei.getAccount(message.caller)
+      let acc = await this.stateManager.getAccount(message.caller)
       if (!acc) {
         acc = new Account()
       }
@@ -971,7 +976,11 @@ export class EVM implements EVMInterface {
     if (account.balance < BigInt(0)) {
       throw new EvmError(ERROR.INSUFFICIENT_BALANCE)
     }
-    const result = this.eei.putAccount(message.authcallOrigin ?? message.caller, account)
+    const result = this.stateManager.putAccount(
+      message.authcallOrigin ?? message.caller,
+      account,
+      true
+    )
     if (this.DEBUG) {
       debug(`Reduced sender (${message.caller}) balance (-> ${account.balance})`)
     }
@@ -985,7 +994,7 @@ export class EVM implements EVMInterface {
     }
     toAccount.balance = newBalance
     // putAccount as the nonce may have changed for contract creation
-    const result = this.eei.putAccount(message.to, toAccount)
+    const result = this.stateManager.putAccount(message.to, toAccount)
     if (this.DEBUG) {
       debug(`Added toAccount (${message.to}) balance (-> ${toAccount.balance})`)
     }
@@ -993,11 +1002,11 @@ export class EVM implements EVMInterface {
   }
 
   protected async _touchAccount(address: Address): Promise<void> {
-    let account = await this.eei.getAccount(address)
+    let account = await this.stateManager.getAccount(address)
     if (!account) {
       account = new Account()
     }
-    return this.eei.putAccount(address, account)
+    return this.stateManager.putAccount(address, account)
   }
 
   /**
@@ -1014,9 +1023,9 @@ export class EVM implements EVMInterface {
     const opts = {
       ...this._optsCached,
       common,
-      eei: this.eei.copy(),
+      stateManager: this.stateManager.copy(),
     }
-    ;(opts.eei as any)._common = common
+    ;(opts.stateManager as any)._common = common
     return new EVM(opts)
   }
 }
