+++ packages/evm/src/index.ts
@@ -3,11 +3,8 @@ import { EvmError, ERROR as EvmErrorMessage } from './exceptions'
 import { InterpreterStep } from './interpreter'
 import { Message } from './message'
 import { getActivePrecompiles } from './precompiles'
-import { VmState } from './state/evmState'
-import { EEI } from './state/state'
 import { EEIInterface, EVMInterface, EVMStateAccess, Log } from './types'
 export {
-  EEI,
   EEIInterface,
   EVM,
   EvmError,
@@ -20,5 +17,4 @@ export {
   InterpreterStep,
   Log,
   Message,
-  VmState,
 }
