+++ packages/evm/src/opcodes/EIP2929.ts
@@ -21,12 +21,12 @@ export function accessAddressEIP2929(
 ): bigint {
   if (common.isActivatedEIP(2929) === false) return BigInt(0)
 
-  const eei = runState.eei
+  const stateManager = runState.stateManager
   const addressStr = address.bytes
 
   // Cold
-  if (!eei.isWarmedAddress(addressStr)) {
-    eei.addWarmedAddress(addressStr)
+  if (!stateManager.isWarmedAddress(addressStr)) {
+    stateManager.addWarmedAddress(addressStr)
 
     // CREATE, CREATE2 opcodes have the address warmed for free.
     // selfdestruct beneficiary address reads are charged an *additional* cold access
@@ -56,13 +56,13 @@ export function accessStorageEIP2929(
 ): bigint {
   if (common.isActivatedEIP(2929) === false) return BigInt(0)
 
-  const eei = runState.eei
+  const stateManager = runState.stateManager
   const address = runState.interpreter.getAddress().bytes
-  const slotIsCold = !eei.isWarmedStorage(address, key)
+  const slotIsCold = !stateManager.isWarmedStorage(address, key)
 
   // Cold (SLOAD and SSTORE)
   if (slotIsCold) {
-    eei.addWarmedStorage(address, key)
+    stateManager.addWarmedStorage(address, key)
     return common.param('gasPrices', 'coldsload')
   } else if (!isSstore) {
     return common.param('gasPrices', 'warmstorageread')
@@ -89,12 +89,12 @@ export function adjustSstoreGasEIP2929(
 ): bigint {
   if (common.isActivatedEIP(2929) === false) return defaultCost
 
-  const eei = runState.eei
+  const stateManager = runState.stateManager
   const address = runState.interpreter.getAddress().bytes
   const warmRead = common.param('gasPrices', 'warmstorageread')
   const coldSload = common.param('gasPrices', 'coldsload')
 
-  if (eei.isWarmedStorage(address, key)) {
+  if (stateManager.isWarmedStorage(address, key)) {
     switch (costName) {
       case 'noop':
         return warmRead
