+++ packages/statemanager/src/stateManager.ts
@@ -1,3 +1,4 @@
+import { Chain, Common, Hardfork } from '@ethereumjs/common'
 import { RLP } from '@ethereumjs/rlp'
 import { Trie } from '@ethereumjs/trie'
 import {
@@ -7,6 +8,7 @@ import {
   KECCAK256_NULL_S,
   KECCAK256_RLP,
   KECCAK256_RLP_S,
+  RIPEMD160_ADDRESS_STRING,
   bigIntToHex,
   bytesToHex,
   bytesToPrefixedHexString,
@@ -15,6 +17,7 @@ import {
   hexStringToBytes,
   setLengthLeft,
   short,
+  toBytes,
   unpadBytes,
   utf8ToBytes,
 } from '@ethereumjs/util'
@@ -23,10 +26,17 @@ import { keccak256 } from 'ethereum-cryptography/keccak'
 import { hexToBytes } from 'ethereum-cryptography/utils'
 
 import { AccountCache, CacheType, StorageCache } from './cache'
+import { Journaling } from './cache/journaling'
 
-import type { AccountFields, StateManagerInterface, StorageDump } from '@ethereumjs/common'
+import type {
+  AccessListItem,
+  AccountFields,
+  StateManagerInterface,
+  StorageDump,
+} from '@ethereumjs/common'
 import type { PrefixedHexString } from '@ethereumjs/util'
 import type { Debugger } from 'debug'
+import type { AccessList } from 'ethers/lib/utils'
 
 export type StorageProof = {
   key: PrefixedHexString
@@ -116,6 +126,11 @@ export interface DefaultStateManagerOpts {
   accountCacheOpts?: CacheOptions
 
   storageCacheOpts?: CacheOptions
+
+  /**
+   * The common to use
+   */
+  common?: Common
 }
 
 /**
@@ -137,9 +152,35 @@ export class DefaultStateManager implements StateManagerInterface {
   _storageTries: { [key: string]: Trie }
   _codeCache: { [key: string]: Uint8Array }
 
+  // EIP-2929 address/storage trackers.
+  // This maps both the accessed accounts and the accessed storage slots.
+  // It is a Map(Address => StorageSlots)
+  // It is possible that the storage slots set is empty. This means that the address is warm.
+  // It is not possible to have an accessed storage slot on a cold address (which is why this structure works)
+  // Each call level tracks their access themselves.
+  // In case of a commit, copy everything if the value does not exist, to the level above
+  // In case of a revert, discard any warm slots.
+  //
+  // TODO: Switch to diff based version similar to _touchedStack
+  // (_accessStorage representing the actual state, separate _accessedStorageStack dictionary
+  // tracking the access diffs per commit)
+  protected _accessedStorage: Map<string, Set<string>>[]
+
+  // Backup structure for address/storage tracker frames on reverts
+  // to also include on access list generation
+  protected _accessedStorageReverted: Map<string, Set<string>>[]
+
+  protected _originalStorageCache: Map<string, Map<string, Uint8Array>>
+
+  protected readonly touchedJournal: Journaling<string>
+
   protected readonly _prefixCodeHashes: boolean
   protected readonly _accountCacheSettings: CacheSettings
   protected readonly _storageCacheSettings: CacheSettings
+
+  protected readonly _common: Common
+
+  protected _checkpointCount: number
 
   /**
    * StateManager is run in DEBUG mode (default: false)
@@ -160,9 +201,19 @@ export class DefaultStateManager implements StateManagerInterface {
 
     this._debug = createDebugLogger('statemanager:statemanager')
 
+    this._common = opts.common ?? new Common({ chain: Chain.Mainnet })
+
+    this._checkpointCount = 0
+
     this._trie = opts.trie ?? new Trie({ useKeyHashing: true })
     this._storageTries = {}
     this._codeCache = {}
+
+    this._originalStorageCache = new Map()
+    this._accessedStorage = [new Map()]
+    this._accessedStorageReverted = [new Map()]
+
+    this.touchedJournal = new Journaling<string>()
 
     this._prefixCodeHashes = opts.prefixCodeHashes ?? true
     this._accountCacheSettings = {
@@ -216,11 +267,26 @@ export class DefaultStateManager implements StateManagerInterface {
   }
 
   /**
+   * Checks if the `account` corresponding to `address`
+   * is empty or non-existent as defined in
+   * EIP-161 (https://eips.ethereum.org/EIPS/eip-161).
+   * @param address - Address to check
+   */
+  async accountIsEmptyOrNonExistent(address: Address): Promise<boolean> {
+    const account = await this.getAccount(address)
+    if (account === undefined || account.isEmpty()) {
+      return true
+    }
+    return false
+  }
+
+  /**
    * Saves an account into state under the provided `address`.
    * @param address - Address under which to store `account`
    * @param account - The account to store or undefined if to be deleted
+   * @param touch - If the account should be touched or not (for state clearing, see TangerineWhistle / SpuriousDragon hardforks)
    */
-  async putAccount(address: Address, account: Account | undefined): Promise<void> {
+  async putAccount(address: Address, account: Account | undefined, touch = false): Promise<void> {
     if (this.DEBUG) {
       this._debug(
         `Save account address=${address} nonce=${account?.nonce} balance=${
@@ -243,6 +309,9 @@ export class DefaultStateManager implements StateManagerInterface {
       } else {
         this._accountCache!.del(address)
       }
+    }
+    if (touch) {
+      this.touchAccount(address)
     }
   }
 
@@ -268,8 +337,9 @@ export class DefaultStateManager implements StateManagerInterface {
   /**
    * Deletes an account from state under the provided `address`.
    * @param address - Address of the account which should be deleted
+   * @param touch - If the account should be touched or not (for state clearing, see TangerineWhistle / SpuriousDragon hardforks)
    */
-  async deleteAccount(address: Address) {
+  async deleteAccount(address: Address, touch = false) {
     if (this.DEBUG) {
       this._debug(`Delete account ${address}`)
     }
@@ -281,6 +351,41 @@ export class DefaultStateManager implements StateManagerInterface {
     if (!this._storageCacheSettings.deactivate) {
       this._storageCache?.clearContractStorage(address)
     }
+    if (touch) {
+      this.touchAccount(address)
+    }
+  }
+
+  /**
+   * Marks an account as touched, according to the definition
+   * in [EIP-158](https://eips.ethereum.org/EIPS/eip-158).
+   * This happens when the account is triggered for a state-changing
+   * event. Touched accounts that are empty will be cleared
+   * at the end of the tx.
+   */
+  touchAccount(address: Address): void {
+    this.touchedJournal.addJournalItem(address.toString().slice(2))
+  }
+
+  /**
+   * Removes accounts form the state trie that have been touched,
+   * as defined in EIP-161 (https://eips.ethereum.org/EIPS/eip-161).
+   */
+  async cleanupTouchedAccounts(): Promise<void> {
+    if (this._common.gteHardfork(Hardfork.SpuriousDragon) === true) {
+      const touchedArray = Array.from(this.touchedJournal.journal)
+      for (const addressHex of touchedArray) {
+        const address = new Address(hexToBytes(addressHex))
+        const empty = await this.accountIsEmptyOrNonExistent(address)
+        if (empty) {
+          await this.deleteAccount(address)
+          if (this.DEBUG) {
+            this._debug(`Cleanup touched account address=${address} (>= SpuriousDragon)`)
+          }
+        }
+      }
+    }
+    this.touchedJournal.clear()
   }
 
   /**
@@ -393,6 +498,47 @@ export class DefaultStateManager implements StateManagerInterface {
   }
 
   /**
+   * Caches the storage value associated with the provided `address` and `key`
+   * on first invocation, and returns the cached (original) value from then
+   * onwards. This is used to get the original value of a storage slot for
+   * computing gas costs according to EIP-1283.
+   * @param address - Address of the account to get the storage for
+   * @param key - Key in the account's storage to get the value for. Must be 32 bytes long.
+   */
+  async getOriginalContractStorage(address: Address, key: Uint8Array): Promise<Uint8Array> {
+    if (key.length !== 32) {
+      throw new Error('Storage key must be 32 bytes long')
+    }
+
+    const addressHex = address.toString()
+    const keyHex = bytesToHex(key)
+
+    let map: Map<string, Uint8Array>
+    if (!this._originalStorageCache.has(addressHex)) {
+      map = new Map()
+      this._originalStorageCache.set(addressHex, map)
+    } else {
+      map = this._originalStorageCache.get(addressHex)!
+    }
+
+    if (map.has(keyHex)) {
+      return map.get(keyHex)!
+    } else {
+      const current = await this.getContractStorage(address, key)
+      map.set(keyHex, current)
+      return current
+    }
+  }
+
+  /**
+   * Clears the original storage cache. Refer to {@link StateManager.getOriginalContractStorage}
+   * for more explanation. Alias of the internal {@link StateManager._clearOriginalStorageCache}
+   */
+  clearOriginalStorageCache(): void {
+    this._originalStorageCache = new Map()
+  }
+
+  /**
    * Modifies the storage trie of an account.
    * @private
    * @param address -  Address of the account whose storage is to be modified
@@ -453,8 +599,14 @@ export class DefaultStateManager implements StateManagerInterface {
    * @param value - Value to set at `key` for account corresponding to `address`.
    * Cannot be more than 32 bytes. Leading zeros are stripped.
    * If it is a empty or filled with zeros, deletes the value.
+   * @param touch - If the account should be touched or not (for state clearing, see TangerineWhistle / SpuriousDragon hardforks)
    */
-  async putContractStorage(address: Address, key: Uint8Array, value: Uint8Array): Promise<void> {
+  async putContractStorage(
+    address: Address,
+    key: Uint8Array,
+    value: Uint8Array,
+    touch = false
+  ): Promise<void> {
     if (key.length !== 32) {
       throw new Error('Storage key must be 32 bytes long')
     }
@@ -475,13 +627,17 @@ export class DefaultStateManager implements StateManagerInterface {
     } else {
       await this._writeContractStorage(address, account, key, value)
     }
+    if (touch) {
+      this.touchAccount(address)
+    }
   }
 
   /**
    * Clears all storage entries for the account corresponding to `address`.
    * @param address -  Address to clear the storage of
+   * @param touch - If the account should be touched or not (for state clearing, see TangerineWhistle / SpuriousDragon hardforks)
    */
-  async clearContractStorage(address: Address): Promise<void> {
+  async clearContractStorage(address: Address, touch = false): Promise<void> {
     let account = await this.getAccount(address)
     if (!account) {
       account = new Account()
@@ -491,6 +647,9 @@ export class DefaultStateManager implements StateManagerInterface {
       storageTrie.root(storageTrie.EMPTY_TRIE_ROOT)
       done()
     })
+    if (touch) {
+      this.touchAccount(address)
+    }
   }
 
   /**
@@ -502,6 +661,11 @@ export class DefaultStateManager implements StateManagerInterface {
     this._trie.checkpoint()
     this._storageCache?.checkpoint()
     this._accountCache?.checkpoint()
+    if (this._common.gteHardfork(Hardfork.Berlin)) {
+      this._accessedStorage.push(new Map())
+    }
+    this._checkpointCount++
+    this.touchedJournal.checkpoint()
   }
 
   /**
@@ -513,6 +677,24 @@ export class DefaultStateManager implements StateManagerInterface {
     await this._trie.commit()
     this._storageCache?.commit()
     this._accountCache?.commit()
+    if (this._common.gteHardfork(Hardfork.Berlin)) {
+      // Copy the contents of the map of the current level to a map higher.
+      const storageMap = this._accessedStorage.pop()
+      if (storageMap) {
+        this._accessedStorageMerge(this._accessedStorage, storageMap)
+      }
+    }
+    this.touchedJournal.commit()
+    this._checkpointCount--
+
+    if (this._checkpointCount === 0) {
+      await this.flush()
+      this.clearOriginalStorageCache()
+    }
+
+    if (this.DEBUG) {
+      this._debug(`state checkpoint committed`)
+    }
   }
 
   /**
@@ -526,6 +708,21 @@ export class DefaultStateManager implements StateManagerInterface {
     this._accountCache?.revert()
     this._storageTries = {}
     this._codeCache = {}
+    if (this._common.gteHardfork(Hardfork.Berlin)) {
+      // setup cache checkpointing
+      const lastItem = this._accessedStorage.pop()
+      if (lastItem) {
+        this._accessedStorageReverted.push(lastItem)
+      }
+    }
+    this.touchedJournal.revert(RIPEMD160_ADDRESS_STRING)
+
+    this._checkpointCount--
+
+    if (this._checkpointCount === 0) {
+      await this.flush()
+      this.clearOriginalStorageCache()
+    }
   }
 
   /**
@@ -731,6 +928,137 @@ export class DefaultStateManager implements StateManagerInterface {
     this._codeCache = {}
   }
 
+  /** EIP-2929 logic
+   * This should only be called from within the EVM
+   */
+
+  /**
+   * Returns true if the address is warm in the current context
+   * @param address - The address (as a Uint8Array) to check
+   */
+  isWarmedAddress(address: Uint8Array): boolean {
+    for (let i = this._accessedStorage.length - 1; i >= 0; i--) {
+      const currentMap = this._accessedStorage[i]
+      if (currentMap.has(bytesToHex(address))) {
+        return true
+      }
+    }
+    return false
+  }
+
+  /**
+   * Add a warm address in the current context
+   * @param address - The address (as a Uint8Array) to check
+   */
+  addWarmedAddress(address: Uint8Array): void {
+    const key = bytesToHex(address)
+    const storageSet = this._accessedStorage[this._accessedStorage.length - 1].get(key)
+    if (!storageSet) {
+      const emptyStorage = new Set<string>()
+      this._accessedStorage[this._accessedStorage.length - 1].set(key, emptyStorage)
+    }
+  }
+
+  /**
+   * Returns true if the slot of the address is warm
+   * @param address - The address (as a Uint8Array) to check
+   * @param slot - The slot (as a Uint8Array) to check
+   */
+  isWarmedStorage(address: Uint8Array, slot: Uint8Array): boolean {
+    const addressKey = bytesToHex(address)
+    const storageKey = bytesToHex(slot)
+
+    for (let i = this._accessedStorage.length - 1; i >= 0; i--) {
+      const currentMap = this._accessedStorage[i]
+      if (currentMap.has(addressKey) && currentMap.get(addressKey)!.has(storageKey)) {
+        return true
+      }
+    }
+
+    return false
+  }
+
+  /**
+   * Mark the storage slot in the address as warm in the current context
+   * @param address - The address (as a Uint8Array) to check
+   * @param slot - The slot (as a Uint8Array) to check
+   */
+  addWarmedStorage(address: Uint8Array, slot: Uint8Array): void {
+    const addressKey = bytesToHex(address)
+    let storageSet = this._accessedStorage[this._accessedStorage.length - 1].get(addressKey)
+    if (!storageSet) {
+      storageSet = new Set()
+      this._accessedStorage[this._accessedStorage.length - 1].set(addressKey, storageSet!)
+    }
+    storageSet!.add(bytesToHex(slot))
+  }
+
+  /**
+   * Clear the warm accounts and storage. To be called after a transaction finished.
+   */
+  clearWarmedAccounts(): void {
+    this._accessedStorage = [new Map()]
+    this._accessedStorageReverted = [new Map()]
+  }
+
+  /**
+   * Generates an EIP-2930 access list
+   *
+   * Note: this method is not yet part of the {@link StateManager} interface.
+   * If not implemented, {@link VM.runTx} is not allowed to be used with the
+   * `reportAccessList` option and will instead throw.
+   *
+   * Note: there is an edge case on accessList generation where an
+   * internal call might revert without an accessList but pass if the
+   * accessList is used for a tx run (so the subsequent behavior might change).
+   * This edge case is not covered by this implementation.
+   *
+   * @param addressesRemoved - List of addresses to be removed from the final list
+   * @param addressesOnlyStorage - List of addresses only to be added in case of present storage slots
+   *
+   * @returns - an [@ethereumjs/tx](https://github.com/ethereumjs/ethereumjs-monorepo/packages/tx) `AccessList`
+   */
+  generateAccessList(
+    addressesRemoved: Address[] = [],
+    addressesOnlyStorage: Address[] = []
+  ): AccessList {
+    // Merge with the reverted storage list
+    const mergedStorage = [...this._accessedStorage, ...this._accessedStorageReverted]
+
+    // Fold merged storage array into one Map
+    while (mergedStorage.length >= 2) {
+      const storageMap = mergedStorage.pop()
+      if (storageMap) {
+        this._accessedStorageMerge(mergedStorage, storageMap)
+      }
+    }
+    const folded = new Map([...mergedStorage[0].entries()].sort())
+
+    // Transfer folded map to final structure
+    const accessList: AccessList = []
+    for (const [addressStr, slots] of folded.entries()) {
+      const address = Address.fromString(`0x${addressStr}`)
+      const check1 = addressesRemoved.find((a) => a.equals(address))
+      const check2 =
+        addressesOnlyStorage.find((a) => a.equals(address)) !== undefined && slots.size === 0
+
+      if (!check1 && !check2) {
+        const storageSlots = Array.from(slots)
+          .map((s) => `0x${s}`)
+          .sort()
+        const accessListItem: AccessListItem = {
+          address: `0x${addressStr}`,
+          storageKeys: storageSlots,
+        }
+        accessList!.push(accessListItem)
+      }
+    }
+
+    return accessList
+  }
+
+  // End of EIP-2929 related logic
+
   /**
    * Dumps the RLP-encoded storage values for an `account` specified by `address`.
    * @param address - The address of the `account` to return storage for
@@ -765,6 +1093,47 @@ export class DefaultStateManager implements StateManagerInterface {
   }
 
   /**
+   * Initializes the provided genesis state into the state trie.
+   * Will error if there are uncommitted checkpoints on the instance.
+   * @param initState address -> balance | [balance, code, storage]
+   */
+  async generateCanonicalGenesis(initState: any): Promise<void> {
+    if (this._checkpointCount !== 0) {
+      throw new Error('Cannot create genesis state with uncommitted checkpoints')
+    }
+    if (this.DEBUG) {
+      this._debug(`Save genesis state into the state trie`)
+    }
+    const addresses = Object.keys(initState)
+    for (const address of addresses) {
+      const addr = Address.fromString(address)
+      const state = initState[address]
+      if (!Array.isArray(state)) {
+        // Prior format: address -> balance
+        const account = Account.fromAccountData({ balance: state })
+        await this.putAccount(addr, account)
+      } else {
+        // New format: address -> [balance, code, storage]
+        const [balance, code, storage, nonce] = state
+        const account = Account.fromAccountData({ balance, nonce })
+        await this.putAccount(addr, account)
+        if (code !== undefined) {
+          await this.putContractCode(addr, toBytes(code))
+        }
+        if (storage !== undefined) {
+          for (const [key, value] of storage) {
+            await this.putContractStorage(addr, toBytes(key), toBytes(value))
+          }
+        }
+      }
+    }
+    await this.flush()
+    // If any empty accounts are put, these should not be marked as touched
+    // (when first tx is ran, this account is deleted when it cleans up the accounts)
+    this.touchedJournal.clear()
+  }
+
+  /**
    * Checks whether there is a state corresponding to a stateRoot
    */
   async hasStateRoot(root: Uint8Array): Promise<boolean> {
@@ -782,6 +1151,30 @@ export class DefaultStateManager implements StateManagerInterface {
       return true
     } else {
       return false
+    }
+  }
+
+  /**
+   * Merges a storage map into the last item of the accessed storage stack
+   */
+  private _accessedStorageMerge(
+    storageList: Map<string, Set<string> | undefined>[],
+    storageMap: Map<string, Set<string>>
+  ) {
+    const mapTarget = storageList[storageList.length - 1]
+
+    if (mapTarget !== undefined) {
+      // Note: storageMap is always defined here per definition (TypeScript cannot infer this)
+      for (const [addressString, slotSet] of storageMap) {
+        const addressExists = mapTarget.get(addressString)
+        if (!addressExists) {
+          mapTarget.set(addressString, new Set())
+        }
+        const storageSet = mapTarget.get(addressString)
+        for (const value of slotSet) {
+          storageSet!.add(value)
+        }
+      }
     }
   }
